changes according to PR commentts
