+++ src/interfaces/index.ts
@@ -44,7 +44,6 @@ export * from './stepper-interfaces/ReviewConfirm/certify-interface'
 export * from './store-interfaces/state-interfaces/account-information-interface'
 export * from './store-interfaces/state-interfaces/auth-information-interface'
 export * from './store-interfaces/state-interfaces/business-interface'
-export * from './store-interfaces/state-interfaces/business-id-interface'
 export * from './store-interfaces/state-interfaces/dissolution-state-interface'
 export * from './store-interfaces/state-interfaces/document-delivery-interface'
 export * from './store-interfaces/state-interfaces/name-request-interface'
