+++ tests/unit/App.spec.ts
@@ -16,6 +16,7 @@ import Stepper from '@/components/common/Stepper.vue'
 import { ConfirmDialog } from '@bcrs-shared-components/confirm-dialog'
 import mockRouter from './MockRouter'
 import Vuelidate from 'vuelidate'
+import { Getter } from 'vuex-class'
 
 // mock fetch() as it is not defined in Jest
 // NB: it should be `global.fetch` but that doesn't work and this does
